pricing: extract reusable profile onboarding steps from WelcomePage
Move the two profile onboarding steps (name/job type, favorite
platforms) and the shared form state/submit logic out of WelcomePage
into components/onboarding, so they can be reused outside the
full-page onboarding layout.
